creates classes for dimo services and simplifies dimo class
